watchtower/conf: remove experimental conditional compilation
